requirements.txt
pillow==10.3.0
google-generativeai==0.5.2
numpy==1.26.4
matplotlib==3.8.4
streamlit==1.33.0
streamlit-drawable-canvas==0.9.3
dtaidistance==2.3.11
pyarrow==15.0.2
ipywidgets==8.1.2